+++ test/runtests.jl
@@ -104,6 +104,26 @@ end
     @test fpp_settings() == "fpp_settings.yaml"
 end
 
+@testset "KiteUtils.jl: New Constructors" begin
+    se1 = Settings("system.yaml")
+    @test se1.sim_settings == "settings.yaml"
+    se2 = Settings("system_ram.yaml")
+    @test se2.model == "data/ram_air_kite_body.obj"
+    se2.model = "hey;)"
+    @test se1.model == "data/kite.obj"
+    @test se2.model == "hey;)"
+    @test se2.foil_file == "data/ram_air_kite_foil.dat"
+    se1.elevation = 420.0
+    se2.elevation = 11.11
+    @test se1.elevation == 420.0
+    @test se2.elevation == 11.11
+    dict1 = se_dict(se1)
+    dict2 = se_dict(se2)
+    dict1["initial"]["elevations"][1] == 420.0
+    dict2["initial"]["elevations"][1] == 11.11
+    
+end
+
 include("test_logfiles.jl")
 include("test_logger.jl")
 include("test_transformations.jl")